+++ src/components/Action/Action.component.tsx
@@ -1,25 +1,15 @@
-import { VoidFn, isFunction, type Nullable } from "@ubloimmo/front-util";
-import { MouseEventHandler, useCallback, useMemo, useState } from "react";
+import { type VoidFn, isFunction, type Nullable } from "@ubloimmo/front-util";
+import { type MouseEventHandler, useCallback, useMemo, useState } from "react";
 import styled from "styled-components";
 
+import { actionContainerStyles } from "./Action.styles";
 import {
-  actionContainerStyles,
-  actionLabelContainerStyles,
-  actionTextWrappingStyles,
-} from "./Action.styles";
-import { Badge, type BadgeProps } from "../Badge";
-import { StaticIcon } from "../StaticIcon";
-import { Text } from "../Text";
-import { Tooltip, TooltipProps } from "../Tooltip";
+  ActionCard,
+  ActionCentered,
+  ActionDefault,
+  ActionLarge,
+} from "./components";
 
-import { FlexLayout, FlexRowLayout, type FlexLayoutProps } from "@layouts";
-import {
-  TextProps,
-  type ColorKeyOrWhite,
-  type PaletteColor,
-  type StyleProps,
-  type TestIdProps,
-} from "@types";
 import {
   isEmptyString,
   useClassName,
@@ -29,27 +19,21 @@ import {
   useTestId,
 } from "@utils";
 
+import type { BadgeProps } from "../Badge";
+import type { TooltipProps } from "../Tooltip";
 import type {
   ActionProps,
-  ActionVariant,
+  ActionStyledProps,
   DefaultActionProps,
+  SizedActionMap,
 } from "./Action.types";
-import type {
-  StaticIconProps,
-  StaticIconSize,
-} from "../StaticIcon/StaticIcon.types";
-
-const staticIconSizeMap: Record<ActionVariant, StaticIconSize> = {
-  default: "s",
-  centered: "s",
-  chunky: "m",
-  card: "m",
-};
+import type { TestIdProps } from "@types";
 
 const defaultActionProps: DefaultActionProps = {
   label: "[Action]",
   icon: "Cursor",
-  variant: "default",
+  size: "default",
+  color: "primary",
   disabled: false,
   badgeLabel: null,
   onClick: null,
@@ -61,9 +45,26 @@ const defaultActionProps: DefaultActionProps = {
 };
 
 /**
+ * Maps Action size variants to their corresponding render components.
+ * Each component handles the specific layout and styling for that size variant.
+ *
+ * @type {SizedActionMap}
+ * @property {React.FC<SizedActionProps>} default - Component for default size variant
+ * @property {React.FC<SizedActionProps>} centered - Component for centered size variant
+ * @property {React.FC<SizedActionProps>} large - Component for large size variant
+ * @property {React.FC<SizedActionProps>} card - Component for card size variant
+ */
+const sizedActionMap: SizedActionMap = {
+  default: ActionDefault,
+  centered: ActionCentered,
+  large: ActionLarge,
+  card: ActionCard,
+};
+
+/**
  * An action button with an icon, label and optional badge
  *
- * @version 0.0.7
+ * @version 0.0.8
  *
  * @param {ActionProps} props - The component's props
  * @returns {JSX.Element}
@@ -72,8 +73,8 @@ const Action = (props: ActionProps & TestIdProps): JSX.Element => {
   const { warn } = useLogger("Action");
 
   const mergedProps = useMergedProps(defaultActionProps, props);
-  const styleProps = useStyleProps(mergedProps);
   const testId = useTestId("action", props);
+  const styleProps = useStyleProps({ ...mergedProps, testId });
   const className = useClassName(props);
 
   if (!props.icon) {
@@ -84,9 +85,9 @@ const Action = (props: ActionProps & TestIdProps): JSX.Element => {
     warn(`Missing required icon, defaulting to ${defaultActionProps.label}`);
   }
 
-  if (props.description && props.variant === "default") {
+  if (props.description && props.size === "default") {
     warn(
-      `Description is not available for default variant. Set variant to "chunky" or "card" to display it.`
+      `Description is not available for default size. Set size to "large" or "card" to display it.`
     );
   }
 
@@ -102,55 +103,18 @@ const Action = (props: ActionProps & TestIdProps): JSX.Element => {
     [mergedProps]
   );
 
-  const staticIconProps = useMemo<StaticIconProps>(() => {
-    const color: ColorKeyOrWhite = mergedProps.disabled
-      ? "white"
-      : isHovering
-      ? "primary"
-      : "gray";
-    const size = staticIconSizeMap[mergedProps.variant];
-    const indicator = mergedProps.indicator;
-    return { size, color, name: mergedProps.icon, indicator };
-  }, [mergedProps, isHovering]);
-
   const badgeProps = useMemo<Nullable<BadgeProps & TestIdProps>>(() => {
-    if (!mergedProps.badgeLabel || isEmptyString(mergedProps.badgeLabel))
+    if (
+      !mergedProps.badgeLabel ||
+      isEmptyString(mergedProps.badgeLabel) ||
+      mergedProps.size === "centered"
+    )
       return null;
     return {
       label: mergedProps.badgeLabel,
-      color: mergedProps.disabled ? "gray" : "primary",
+      color: mergedProps.disabled ? "gray" : mergedProps.color,
       shade: "light",
       testId: "action-badge",
-    };
-  }, [mergedProps]);
-
-  const textProps = useMemo<TextProps>(() => {
-    const color: PaletteColor = mergedProps.disabled
-      ? "gray-600"
-      : isHovering
-      ? "primary-base"
-      : "gray-800";
-    return {
-      color,
-      weight: "bold",
-      size: "m",
-    };
-  }, [mergedProps, isHovering]);
-
-  const layoutProps = useMemo<FlexLayoutProps>(() => {
-    if (mergedProps.variant === "chunky" || mergedProps.variant === "card") {
-      return {
-        direction: "column",
-        gap: "s-1",
-        align: "start",
-        justify: "center",
-      };
-    }
-    return {
-      direction: "row",
-      gap: "s-2",
-      align: "center",
-      justify: "space-between",
     };
   }, [mergedProps]);
 
@@ -166,12 +130,15 @@ const Action = (props: ActionProps & TestIdProps): JSX.Element => {
     };
   }, [mergedProps]);
 
-  const canDisplayDescription = useMemo(() => {
-    return (
-      mergedProps.description &&
-      (mergedProps.variant === "chunky" || mergedProps.variant === "card")
-    );
-  }, [mergedProps]);
+  const SizedAction = useMemo(() => {
+    if (!(mergedProps.size in sizedActionMap)) {
+      warn(
+        `Invalid size: ${mergedProps.size} provided, defaulting to "default"`
+      );
+      return sizedActionMap.default;
+    }
+    return sizedActionMap[mergedProps.size];
+  }, [mergedProps.size, warn]);
 
   return (
     <ActionContainer
@@ -185,86 +152,14 @@ const Action = (props: ActionProps & TestIdProps): JSX.Element => {
       className={className}
       {...styleProps}
     >
-      {mergedProps.variant === "centered" ? (
-        <div>
-          <StaticIcon {...staticIconProps} />
-          <Text {...textProps} testId={`${testId}-label`}>
-            {mergedProps.label}
-          </Text>
-        </div>
-      ) : mergedProps.variant === "card" ? (
-        <FlexLayout direction="column" gap="s-2" fill>
-          <StaticIcon {...staticIconProps} />
-          <FlexLayout {...layoutProps} fill>
-            <ActionLabelContainer
-              align="center"
-              justify="start"
-              fill
-              gap="s-2"
-              testId={`${testId}-label-container`}
-            >
-              <Text {...textProps} testId={`${testId}-label`}>
-                {mergedProps.label}
-              </Text>
-              {iconTooltipProps && <Tooltip {...iconTooltipProps} />}
-              {canDisplayDescription && badgeProps && (
-                <Badge {...badgeProps} testId={`${testId}-badge`} />
-              )}
-            </ActionLabelContainer>
-
-            {!canDisplayDescription && badgeProps && (
-              <Badge {...badgeProps} testId={`${testId}-badge`} />
-            )}
-
-            {canDisplayDescription && (
-              <ActionDescription
-                color="gray-600"
-                fill
-                size="s"
-                testId={`${testId}-description`}
-              >
-                {mergedProps.description}
-              </ActionDescription>
-            )}
-          </FlexLayout>
-        </FlexLayout>
-      ) : (
-        <>
-          <StaticIcon {...staticIconProps} />
-          <FlexLayout {...layoutProps} fill>
-            <ActionLabelContainer
-              align="center"
-              justify="start"
-              fill
-              gap="s-2"
-              testId={`${testId}-label-container`}
-            >
-              <Text {...textProps} testId={`${testId}-label`}>
-                {mergedProps.label}
-              </Text>
-              {iconTooltipProps && <Tooltip {...iconTooltipProps} />}
-              {canDisplayDescription && badgeProps && (
-                <Badge {...badgeProps} testId={`${testId}-badge`} />
-              )}
-            </ActionLabelContainer>
-
-            {!canDisplayDescription && badgeProps && (
-              <Badge {...badgeProps} testId={`${testId}-badge`} />
-            )}
-
-            {canDisplayDescription && (
-              <ActionDescription
-                color="gray-600"
-                fill
-                size="s"
-                testId={`${testId}-description`}
-              >
-                {mergedProps.description}
-              </ActionDescription>
-            )}
-          </FlexLayout>
-        </>
-      )}
+      <SizedAction
+        {...mergedProps}
+        isHovering={isHovering}
+        testId={testId}
+        onClick={onClick}
+        badgeProps={badgeProps}
+        iconTooltipProps={iconTooltipProps}
+      />
     </ActionContainer>
   );
 };
@@ -273,14 +168,6 @@ Action.defaultProps = defaultActionProps;
 
 export { Action };
 
-const ActionContainer = styled.button<StyleProps<DefaultActionProps>>`
+const ActionContainer = styled.button<ActionStyledProps>`
   ${actionContainerStyles}
-`;
-
-const ActionLabelContainer = styled(FlexRowLayout)`
-  ${actionLabelContainerStyles}
-`;
-
-const ActionDescription = styled(Text)`
-  ${actionTextWrappingStyles}
 `;